requirements.txt
google-api-python-client 
google-auth
google-auth-oauthlib
google-auth-httplib2
opencv-python
moviepy
ultralytics
numpy
librosa
pillow
requests
garminconnect
fitparse
pandas